backend/requirements.txt
# ===========================================
# Workflow AI 工作流系统依赖
# ===========================================

# -------------------------------------------
# Web 框架 (必需)
# -------------------------------------------
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
websockets==12.0

# -------------------------------------------
# 数据库 (必需)
# -------------------------------------------
sqlalchemy==2.0.25
alembic==1.13.1

# -------------------------------------------
# 数据验证与配置 (必需)
# -------------------------------------------
pydantic>=2.7.4
pydantic-settings==2.1.0

# -------------------------------------------
# 工作流引擎 (必需)
# -------------------------------------------
langgraph==0.0.50
langchain==0.1.8
langchain-openai==0.0.6

# -------------------------------------------
# 环境变量 (必需)
# -------------------------------------------
python-dotenv==1.0.0

# -------------------------------------------
# 知识库 / RAG 依赖 (可选，如果需要知识库功能)
# -------------------------------------------
pymilvus>=2.3.0          # Milvus 向量数据库客户端
pypdf>=4.0.0             # PDF 文件解析
requests>=2.31.0         # HTTP 请求（用于调用 BGE 服务）

# ===========================================
# 安装命令:
#   pip install -r requirements.txt
#
# 注意: 确保 Python 版本 >= 3.9
# ===========================================
